updated the structure of the homeview and added postfeed component, added tailwind-scrollbar plugin
