+++ Frontend/package.json
@@ -22,7 +22,7 @@
     "@tailwindcss/forms": "^0.5.7",
     "@tsconfig/node20": "^20.1.2",
     "@types/jsdom": "^21.1.6",
-    "@types/node": "^20.11.10",
+    "@types/node": "^20.11.19",
     "@vitejs/plugin-vue": "^5.0.3",
     "@vue/eslint-config-typescript": "^12.0.0",
     "@vue/test-utils": "^2.4.4",
@@ -33,6 +33,7 @@
     "jsdom": "^24.0.0",
     "npm-run-all2": "^6.1.1",
     "postcss": "^8.4.34",
+    "tailwind-scrollbar": "^3.1.0",
     "tailwindcss": "^3.4.1",
     "typescript": "~5.3.0",
     "vite": "^5.0.11",
