+++ Frontend/src/style.css
@@ -1,3 +1,9 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+@layer components {
+    .scrollbar-styled {
+        @apply scrollbar-thin scrollbar-thumb-[#3c4345] scrollbar-track-transparent
+    }
+}
