+++ Frontend/tailwind.config.ts
@@ -16,7 +16,7 @@ export default {
     },
   },
   plugins: [
-
+      require('tailwind-scrollbar')
   ],
 } satisfies Config
